requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Werkzeug==3.0.3
python-dotenv==1.0.1
email-validator==2.1.1
